Modified path removal for srcFiles
I was having issues when the current working directory wasn't the same as the full src path to the images - this was resulting in some of the path remaining in the srcFiles array, causing the difference check to fail. I've modified the srcFiles to remove the path based of the file.base value. 

This seems to be working correctly for me, but it would be worth checking that this works as was originally intended before merging back in.